refactor(utils.projen-builder): extract `OptionsPropertyBuilder/`NameSchemeBuilder` to builders
